Change to timestamp-based versioning.
Now when the Makefile runs for the default target, it will
always recompile version.c, providing it with a defined symbol
BUILD_TIME corresponding to the current time as YYMMDDHHMMSS.
This string is displayed on the About screen and on the
corresponding Terminal response in place of the hard-coded
version number shown before.

A new script version.sh displays the BUILD_TIME used for the
most recent build, for comparison with the displayed value.

The Makefile change is perhaps not completely integrated
with the automation in the Nordic SDK's build system, but
it should work for our purposes.