Make the I2C utilities work (detects devices)
Check whether the I2C bus was successfully initialized before
trying to use it, in each low-level routine.

Enhance the I2C feature on the test screen to diagnose when the
I2C bus is not working at all (stuck low).

Swap the I2C SCL and SDA pins to match the hardware.

Change the dummy clock pin associated with SPI2 from 22 to 29, to
avoid a conflict with the pin we chose for SCL. Pin 29 is unused on
Bender, JoCo, and Trans-Ionospheric.

Also, fix the Makefile machinery for generating the timestamp to work
even right after a make clean.

At this point the I2C routines successfully detect an I2C device
and update the test screen display.