Add i2c_write command to the TCLish interpreter
i2c_write seven-bit-address bytecount byte0 byte1 ...

This is the first command that requires a variable number of arguments,
so a minor change to the interpreter innards was also necessary. Each
command knows its "arity", the number of items in the command, and this
value is checked before calling the function. Now, the arity can be
negative, and that means the number of items must be AT LEAST -arity.