New top-level graphics to suit Trans-Ionospheric theme
JINTRO.RAW plays after the start-up text messages and before the
main menu is presented. This one is an animation of the front
panel of an actual Zenith Trans-Oceanic model G500, with the
tuning indicator sweeping from one end of the dial to the other,
and back. (original photography)

INTRO.RAW (deleted here) was the same thing for Bender.

BLING/TRANSIO/MAGICEYE.RAW is an animation of a magic eye tube,
based on https://commons.wikimedia.org/wiki/File:Em11-ani.gif
This is intended to be used as the image shown after exiting
the main menu. Bender called this "unlock mode" because the
LEDs show which unlocks have been activated on the badge.

MENU/DEFAULT.PRV is the background image shown behind menus when
there's no menu item preview to show. This one is a close-up of
the crowned-shield logo of the Zenith Trans-Oceanic G500.
(original photography)

Hint: .RAW files can be played on a computer like this:

ffplay -f rawvideo -s 128x128 -pix_fmt rgb565be FILENAME.RAW

and .PRV files can be played on a computer like this:

ffplay -f rawvideo -s 101x112 -pix_fmt rgb565be FILENAME.PRV