Remove tooth- and eye-specific LED display processing
Rename all the tooth and eye functions which updated LEDs in the
background when nothing else was using them with generic names.

Note that Bender was able to update the eyes and cigar independent
of the other LEDs because they were first in the serial data chain.
Thus the terminal commands "led" and "leds" were able to change the
matrix of 12 LEDs in Bender's mouth even when the background bling
was running. JoCo's background processing updated all the LEDs, so
the terminal commands were broken most of the time. That's still
the case for Trans-Ionospheric, as of this commit.