Implement I2C bargraph driver, enhance I/O expander
I2C bargraph add-on can now be started, stopped, and updated with a
value in the range 0-24, where 0 is black and 24 is all bars lit. Each
bar on the bargraph has a predefined color, chosen to resemble a common
S-meter configuration.

The system test screen feature for I2C test is enhanced to operate the
bargraph from the same counter as the LEDs on the I2C I/O expander
add-on.

The I/O expander driver is a little better encapsulated, and enhanced
to operate all 16 bits on the device as outputs.

The I/O expander test now also uses all 16 bits. The upper byte counts
down while the lower byte counts up.