WIP: Implement structure of Mastermind codebreaker UI
New concept: each Mastermind game is played in both directions at once.
One player chooses Mastermind from the Games menu, and gets to choose
their opponent (either a remote badge or the built-in automatic player).
In case of a remote badge, the other player gets invited to play via a
pop-up. Each player starts by creating a secret code for the other
player to guess. Both players then start guessing. The guesses are sent
over to the opposite badge, where the codemaker's function of providing
answers is handled automatically with no user intervention.

The user interface is mostly graphical. Pegs are displayed as colored
boxes on the screen. User entry isn't implemented as of this commit,
but it will work like this: a moveable arrow is displayed below the
row of boxes where the entry is to be used. Beside the row of boxes,
two other menu items appear: Go and Quit. Button function:

  UP        change the box above the arrow to the next color
  DOWN      change the box above the arrow to the previous color
  LEFT      move the arrow to the left (NOT cancel as usual on the badge)
  RIGHT     move the arrow to the right
  ACTION    select the menu item above the arrow (if enabled)

The UI starts out with nice big boxes on the screen, but there's only
room for six rows at that size. On the seventh turn, the boxes rescale
to make room for twelve rows, which is the standard turn limit for
Mastermind.

There's a cute little animation of a dialup modem to fill the time
while a remote connection is being processed.

Lots of this is still stubbed out. Look for "//!!!" flagging places
that need implementing.