Integrate Mastermind scorer; first playable version
Integrated Michelle's code for scoring the guesses. It was
written for the console environment independent of my UI
development, so I had to add some adapter code at the
beginning and end (and comment out the printf calls).

There's a little timer-driven animation while we wait for the
score to arrive. That delay is totally fake for local scoring,
of course, but it will fill time during any network delays for
remote play.

Switched a peg color from orange to magenta because orange was
too close visually to yellow on the LCD.

Added the routine where the local user enters the secret code
for the remote user to guess. This required some changes to
the guess entry routines to make them more general.

Victory detection has now been tested a little, resulting in
a minor fix to the congratulatory message.

The auto opponent now generates a random secret code. This is
done in the "connect" routine.