Block events during a Mastermind game (etc)
The drawing code I wrote for Mastermind isn't set up to be interrupted
and then to redraw the screen after an event like a Hello or a Score
comes around and takes over the screen. For now, we just pause the
event scheduler so that no such events can happen.

Also added a hint on screen explaining that an empty black box is
not, in fact, a color choice. This hint comes up on the secret code
picker screen, which is the first thing the local users sees. We
assume he still remembers it on the main game screen.