Add hint for key peg interpretation
On the main game screen, the first time (in each game) that the user
sees an answer, the bottom of the screen is filled with an answer key
explaining the color code.

This can be a little confusing if the user gets an empty answer.

Also changed the key pegs for exact match from pink to red. I just
like it better that way. Even better would be a color that's very
distinct from any color on the guess side, but we're out of colors.