Add a Callsign setting to the badge state
The callsign setting is prompted for (and required) before entering
the log viewer. Otherwise it remains unset.

The callsign length is limited to 6 characters. That is enough to
handle any standard amateur callsign without prefixes or suffixes.
That wouldn't be enough for a real logging program, but hey!

Also abstracted the log viewer a bit, to make it potentially easier
for other code to use it as a generic file viewer.

Looks like I failed to include the transio_log.* files in the
previous commit, so it wouldn't build.