Update HELLO for Trans-Ionospheric
Trans-Ionospheric badges now get a distinctive Hello.

It's based on the startup animation. I grab just the first frame
of the startup animation, because it was quicker than making a
single-frame .RAW file and updating all the SD cards.

I didn't like the legibility of the drop-shadow text, so I
changed the "shadow" to a filled rectangle behind the text.

I also didn't like waiting 60 seconds for a hello, so I
changed the minimum hello interval to 20 seconds.

In lieu of actually updating the neighbor list processing,
Trans-Ionospheric badges are now treated like JoCos for list
processing purposes.