Add basic support for running GDB using BlackMagic Probe
There's a new Makefile target, gdb, which includes all the magic
words required to run it, together with a startup script.

It does assume that the BlackMagic Probe is the only (or first)
thing in the system that creates serial ports named /dev/ttyACM*.
To use a port other than /dev/ttyACM0, modify the startup script
firmware/gdb-init.