Removed Active List processing
The Active List was for JoCo "visit" detection, and meant that badges
we'd already visited were not shown as nearby. We don't want that
behavior.

Also added processing for both the official Open Research Institute
company ID and the temporary one we used at Hamvention. Both are
handled identically for now.