Checkpoint on Neighbor List rewrite
Code compiles but there are significant pieces still unwritten.
Everything is completely UNTESTED at time of commit.

See the new design note in firmware/notes/new-database-scheme.txt for a
discussion of what these code changes are about in some detail.

Changes so far:

The menu system is enhanced to support dynamic submenus. In short, a
submenu can now support an unlimited number of entries without
allocating a static array to hold them all. What's more, the entries can
be updated in real time while they are displayed on the screen.

The Nearby menu is partially converted over to use the dynamic menu
system.

Button wait is enhanced to support a timeout, so things can happen
between button presses. This is to let us update RSSI and Name on the
Nearby menu.

The neighbor list maintenance in util_ble_lists is totally revamped. See
the design note. I tried to minimize complexity and also minimize
processing load for each received advertisement.

BLE appearance codes are checked. Code 0x19DC is interpreted according
to last year's rules, and code 0x26DC is interpreted according to the
written badgestdcom spec.

All of the manufacturer ID codes that have been published in the
Badgelife slack are now interpreted and used to provide default names
for those badges. However, if the badge provides a name in the Bluetooth
standard way (with the entire name in the advertisement) the transmitted
name is used for all badge types, even unknown ones.

The size of badge names is increased from 8 characters (9 bytes) to 10
characters (11 bytes). This is to comply with the written badgestdcom
specification, even though it appears AND!XOR still use 8-character
names.

Device IDs are defaulted to the last two bytes of the GAP address,
except for DC25 badges (which transmit device ID in the first and second
bytes of the company-specific data) and DC26 AND!XOR badges (which
transmit device ID in the second and third bytes of the company-specific
data).

Removed support for Medea vodka bottles. I don't have one to test with,
and the new architecture for the neighbor list means they won't fit in
as before.

The entire Mastermind game is commented out for this build, because it
needs to be converted over to use the new neighbor list system.

Default name for the badge changed from ELMER to Phase4, since it was
pointed out to me that ELMER has sexist overtones.

Still totally unwritten, though sketched out in design notes:

__nearby_callback() that implements the action on selecting a neighbor.

ble_lists_draw_callback() that draws an entry on the neighbor list
screen.

ble_lists_process_advertisement() that processes a received
advertisement according to the new scheme.

Modifications to Mastermind to use the new scheme.

Probably other things.