Display Info on selected neighbor list badge
When you select an entry on the neighbor list, we now pop up
an info screen. In the future this will be sensitive to whether
the badge is playing the QSO game, but for now it's just a plain
info screen about the neighbor badge.

Unfortunately added company_id to the neighbor list. We can't let
this struct get too big, but I really wanted to display the
badgemaker name on the info screen. Could encode it in one byte
instead of two, but didn't.

Encapsulated the short names corresponding to company_id values
into a function so we can use it on the info screen too.

Changed the font for pop-up titles from LARGE to MEDIUM because
the LARGE font is all caps and I want to preserve lowercase in
badge names.