Brought back SCORE screen, revised scoring.
Revised the SCORE screen to show QSO count and points, but not level, and
remove reference to trinkets.

Removed the provision to set the high bit of the BLE claimed score when
the badge is eligible for a trinket.

Defined scoring for QSO attempts and QSOs completed.

Removed most support for game levels. Added support for QSO Count.

Now it's become obvious that I repurposed the SCORE icon on the main screen
for the VIEWLOG function, so we're going to need a new icon for that.