Checkpoint: First QSO from neighbor list
We now have enough mechanism to actually initiate and complete a
basic QSO exchange with another badge from the neighbor list.
If we select a neighbor list entry and hit the Action button,
if it's a QSO-game enabled badge, we get a dialog that lets us
choose to initiate the QSO.

Advertisement now includes the QSO game flag. It's always set iff the
incoming games bit is set.

gamefile code (convert BLE address to filename) integrated (not called yet).