Tweaks to QSO game, phase 1 of N
Better management of the game flags. Default correctly (I hope)
at poweron, and set correctly when circumstances change.

Started code to animate the LEDs during a QSO connect, and make
it abortable with the keyboard, but that's incomplete still.

Re-ordered the messages on the QSO screen to make more sense.