Re-integrate Mastermind into the system. No remote yet.
Brought the Mastermind game back into the build.

Playing Mastermind now generates points for the score. Solving the
puzzle fast get the most points. Solving it slower gets lots of points,
too. Failing to solve it gets a consolation prize of points. Quitting
or having a network failure makes no points.

Mastermind can now select a remote opponent from the neighbor list
and obtain its Bluetooth GAP address. Actually connecting to the
remote opponent and playing the game over the air remains to be
implemented later.

To implement the remote players list, the neighbor list can now
be filtered.

As a consequence of integration with the neighbor
list, the UI logic is slightly changed. The user has to answer a
separate question to decide whether to play against the badge when
there are remote players present, instead of "Play the Badge" being
one of the menu choices along with the remote players.

During the screen animation that fills time during a Bluetooth remote
connection (fake for now), instead of 867-5309 we "dial" the GAP address.

Added Mastermind puzzles solved to the score screen.

Made qso_count 16 bits instead of 8; it's easy to make 256 QSOs.

Added a cheat code for Mastermind to make it easier to debug the scoring.
When the cheat code is enabled, the answer pegs are not scrambled. This
makes it much easier to solve the puzzles.

The BLE advertisement flags now implement the Mastermind bit.