New HELLO handling, phase 1
Hellos are now synchronized with the background LED sequence. When the
radio-control simulation ends, there's an opportunity for a Hello. At
each opportunity, the neighbor list is scanned for the best Hello
candidate.

That means THERE IS NO TRY to Hello.

All of the event handlers for Hello blings have been consolidated into a
single function based on the Trans-Ionospheric handler. All are able to
display the HELLO BADGENAME on top of an image, which doesn't have to be
dimmed to work acceptably well. All use the same LED sequence, which
lights up in a random color and then incrementally fades out.

As of this commit, the neighbor list scan is dummied out. That's next.