Debug publication of bragtape characteristic
Buffer limit for static sized buffers is 510, not 512.

Adding characteristics on the fly is problematic; can't generally
be done because the SDK requires characteristics to be added in
sequence. We could probably have still deferred adding the brag
characteristic, but it wouldn't scale and there doesn't seem to
be any way to delete a characteristic. Just gave up and allocated
the buffer and the characteristic at startup. Receiving badge
will have to detect an empty brag message as no message.