Add reception of bragtape to QSO game
Fetches the bragtape characteristic from the remote badge during a
QSO. Notifies the user of the received message, but does not yet
display it inline.

Adds the message to the log file. Log file viewer slightly enhanced
to display non-printable-ASCII characters as '_'.

Log file viewer currently truncates lines at the screen width.
Needs to wrap lines instead.

Characteristic fetching is implausibly slow. Need to find out why.
If it's going to be that slow, the message needs to display as
it is received.